Haskollider.Buffer

Description

Haskollider.Buffer module contains the set of functions for buffer control on the scsynth server

Documentation

data Buffer

Buffer type that contains the number of frames, number of channels, number of the buffer itself, and message used to instance it.

Constructors

Buffer

Fields

numFrames :: Int

numChannels :: Int

bufnum :: Int

bMessage :: Message

alloc :: Int -> Int -> Int -> Buffer

Allocate buffer message, takes a number of frames and channels, and returns a function that when given a number will return a Buffer object

allocRead :: String -> Int -> Int -> Int -> Buffer

Similar to alloc, but takes a path to an audio file to read into the buffer

allocReadChannel :: String -> Int -> Int -> Int -> Int -> Buffer

Similar to allocRead, but reads a specified number of channels

query :: Buffer -> Message

read :: Buffer -> String -> Int -> Int -> Bool -> Message

readChannel :: Buffer -> String -> Int -> Int -> Bool -> Int -> Message

write :: Buffer -> String -> String -> String -> Int -> Int -> Bool -> Message

freeBuf :: Buffer -> Message

close :: Buffer -> Message

sendCollection :: Buffer -> [Double] -> Int -> Maybe [Message]

Returns an osc message for sending a collection of values to a buffer on the server

sinList :: Double -> [Double]

messageListToDoubleList :: Maybe [Message] -> Maybe [Double]

testSendCollection :: Int -> IO Bool

Test for our sendCollection function. It looks like we've got it working right!