Haskollider.Bus

Description

Haskollider.Bus defines the OSC message composition functions for controlling busses in scsynth

Documentation

data Rate

Bus rate algebraic type. Audio rate has a faster update rate.

Constructors

Control

Audio

data Bus

Bus type for controlling buses in scsynth. Takes a rate, index, and number of channels

Constructors

Bus

Fields

rate :: Rate

index :: Int

numChannels :: Int

isSettable :: Bus -> Bool

set :: Bus -> Double -> Message

setn :: Bus -> [Double] -> Message

setAt :: Bus -> Int -> [Double] -> Message

fill :: Bus -> Double -> Int -> Message

setAll :: Bus -> Double -> Message

setValue :: Bus -> Double -> Message

newBus :: Rate -> Int -> Int -> Bus

Given a rate and number of channels, returns a function that takes an index and returns a bus. Use with sendBus