Haskollider.Engine

Description

Engine module contains internal methods for node, buffer, and bus allocation

Documentation

type NodeID = Int

initialNodeID :: NodeID

Default initial node

allocNodeID :: NodeID -> NodeID

Given a previous node id, returns the next available node.

resetNodeID :: NodeID

Returns the initialNodeID, 1000

data PowerOfTwoBlock

Constructors

PowerOfTwoBlock

Fields

blockAddress :: Int

blockSize :: Int

blockNext :: Maybe PowerOfTwoBlock

type BlockRegion = Vector (Maybe PowerOfTwoBlock)

replaceBlockAt :: Int -> Maybe PowerOfTwoBlock -> BlockRegion -> BlockRegion

data PowerOfTwoAllocator

Constructors

PowerOfTwoAllocator

Fields

allocatorSize :: Int

allocatorArray :: BlockRegion

allocatorFreeList :: BlockRegion

allocatorPos :: Int

newPowerOfTwoAllocator :: Int -> PowerOfTwoAllocator

allocPTBlock :: PowerOfTwoAllocator -> Int -> (Maybe NodeID, PowerOfTwoAllocator)

freePTBlock :: PowerOfTwoAllocator -> Int -> PowerOfTwoAllocator

potAllocUnitTest :: IO ()

Test for the PowerOfTwoAllocator to check correct behavior.