Haskollider.Node

Description

Haskollider.Node defines a set of functions for Osc message composition that allows for the creation and control of Nodes, Synths, and Groups in scsynth

Documentation

data AddAction

AddAction is a simple enum used to organize nodes upon creation

Constructors

AddToHead

AddToTail

AddBefore

AddAfter

AddReplace

type ControlPair = (String, Double)

Control pair is a tuple containing an argument name and value for synths

type ControlList = [ControlPair]

ControlList is a list of ControlPairs, used for multipled updates

rootNodeID :: NodeID

NodeID for the root node on the server

defaultGroupID :: NodeID

NodeID for the default group on the server

grainNodeID :: NodeID

NodeID for grain nodes. -1 is used since they won't be updated after creation.

controlToDatum :: ControlList -> [Datum]

Internal function for converting a ControlList to a list of Osc Datum

class Node a where

Node is the abstract base class for groups and synths. Many functions are useful to any subclass such as free and moveToHead

Methods

nodeID :: a -> NodeID

freeNode :: Node a => a -> Message

run :: Node a => a -> Bool -> Message

set :: Node a => a -> String -> Double -> Message

setList :: Node a => a -> ControlList -> Message

setn :: Node a => a -> Int -> Int -> [Double] -> Message

fill :: Node a => a -> Int -> Int -> Double -> Message

release :: Node a => a -> Double -> Message

trace :: Node a => a -> Message

moveBefore :: Node a => a -> a -> Message

moveAfter :: Node a => a -> a -> Message

moveToHead :: Node a => a -> Group -> Message

moveToTail :: Node a => a -> Group -> Message

data Group

Group is the node subclass that defines an interface for organizing contained nodes

Constructors

Group NodeID Message

moveNodeToHead :: Node n => Group -> n -> Message

moveNodeToTail :: Node n => Group -> n -> Message

freeAll :: Group -> Message

deepFree :: Group -> Message

dumpTree :: Group -> Message

newGroup :: NodeID -> Group

A curried function that takes a NodeID and returns a Group object. Use in conjunction with sendGroup

newGroup' :: Node a => a -> AddAction -> NodeID -> Group

Given a target node and AddAction, returns a curried functions that given a NodeID will return a Group

groupAfter :: Node a => a -> NodeID -> Group

groupBefore :: Node a => a -> NodeID -> Group

groupHead :: Group -> NodeID -> Group

groupTail :: Group -> NodeID -> Group

groupReplace :: Node a => a -> NodeID -> Group

data Synth

Synth is a Node subclass for nodes that actual generate audio and control output in scsynth

Constructors

Synth String NodeID Message

synthName :: Synth -> String

newSynth :: String -> ControlList -> NodeID -> Synth

Takes a SynthDef name and list of arguments/values, returns a curried function. Use with sendSynth.

newSynth' :: Node a => String -> ControlList -> a -> AddAction -> NodeID -> Synth

Takes a SynthDef name, list of arguments/values, target, and AddAction, returns a curried function. Use with sendSynth.

synthAfter :: Node a => a -> String -> ControlList -> NodeID -> Synth

synthBefore :: Node a => a -> String -> ControlList -> NodeID -> Synth

synthHead :: Group -> String -> ControlList -> NodeID -> Synth

synthTail :: Group -> String -> ControlList -> NodeID -> Synth

synthReplace :: Node a => a -> String -> ControlList -> NodeID -> Synth

grainSynth :: String -> ControlList -> Message

Given a name and control list, returns an OscMessage for a grain synth. Use with send.

grainSynth' :: Node a => String -> ControlList -> a -> AddAction -> Message

Given a name, control list, target, and AddAction, returns an OscMessage for a grain synth. Use with send.