Haskollider.Server

Description

Haskollider.Server defines all the types and functionals relavent to communicating with a running scsynth instances

Documentation

data Protocol

Protocol to use for Osc communication with scsynth. Udp is the default for server communication.

Constructors

Tcp

Udp

data ServerType

ServerType, distinguishes between an internal server or local host scsynth instance. LocalHost by default

Constructors

LocalHost

InternalServer

data ServerProcess

ServerProcess is just binds together the various handles returned from createProcess

Constructors

ServerProcess

Fields

stdIn :: Handle

stdOut :: Handle

stdErr :: Handle

process :: ProcessHandle

data ServerOptions

ServerOptions, a group of options passed to the server on creation.

Constructors

ServerOptions

Fields

sNumAudioBusChannels :: Int

sNumControlBusChannels :: Int

sMaxLogins :: Int

sMaxNodes :: Int

sNumInputBusChannels :: Int

sNumOutputBusChannels :: Int

sNumBuffers :: Int

sMaxSynthDefs :: Int

sProtocol :: Protocol

sBufLength :: Int

sNumRGens :: Int

sMaxWireBufs :: Int

sPreferredSampleRate :: Int

sLoadGraphDefs :: Bool

sVerbosity :: Int

sRendezvous :: Bool

sRemoteControlVolume :: Bool

sMemoryLocking :: Bool

sPreferredHardwareBufferFrameSize :: Int

sRealTimeMemorySize :: Int

sBlockSize :: Int

sPortNum :: Int

sNumPrivateAudioBusChannels :: Int

defaultServerOptions :: ServerOptions

A default collection of ServerOptions for quick use and reference. If you want to target specific arguments start here and change via record syntax.

data Server

Server, the main type used to instantiate and communicate with an scsynth instance

Constructors

Server

Fields

serverName :: String

serverOptions :: ServerOptions

serverNetAddr :: NetAddr

serverClientID :: Int

serverRunning :: Bool

serverBootNotifyFirst :: Bool

serverLatency :: Double

serverDumpMode :: Int

serverNotify :: Bool

serverNotified :: Bool

serverNumUgens :: Int

serverNumSynths :: Int

serverAlive :: Bool

serverAliveThreadPeriod :: Double

serverRecHeaderFormat :: String

serverRecSampleFormat :: String

serverRecChannels :: Int

currentNodeID :: Int

controlBusAllocator :: PowerOfTwoAllocator

audioBusAllocator :: PowerOfTwoAllocator

bufferAllocator :: PowerOfTwoAllocator

rootNode :: Group

defaultGroup :: Group

recordBuf :: Maybe Buffer

recordNode :: Maybe Synth

serverCmd :: String

serverProcess :: Maybe ServerProcess

defaultServer :: Server

A Server with typical values.

data ServerCommand

Constructors

None

Notify

Status

Quit

Cmd

D_recv

D_load

D_loadDir

D_freeAll

S_new

N_trace

N_free

N_run

N_cmd

N_map

N_set

N_setn

N_fill

N_before

N_after

U_cmd

G_new

G_head

G_tail

G_freeAll

C_set

C_setn

C_fill

B_alloc

B_allocRead

B_read

B_write

B_free

B_close

B_zero

B_set

B_setn

B_fill

B_gen

DumpOSC

C_get

C_getn

B_get

B_getn

S_get

S_getn

N_query

B_query

N_mapn

S_noid

G_deepFree

ClearSched

Sync

D_free

B_allocReadChannel

B_readChannel

G_dumpTree

G_queryTree

Error

S_newargs

N_mapa

N_mapan

N_order

NUMBER_OF_COMMANDS

firstPrivateBus :: ServerOptions -> Int

Helper function to find the first private bus

optionsToCmdString :: ServerOptions -> [String]

Converts ServerOptions to a list of command line Arguments for scsynth

sendMsg :: Server -> Message -> IO ()

Helper function for sending OSC messages to scsynth

sendBundle :: Server -> Bundle -> IO ()

Helper function for sending OSC bundles to scsynth

nextNodeID :: StateT Server IO NodeID

Increments the server's node counter, returning a node id.

allocControlBus :: Int -> StateT Server IO (Maybe Int)

Allocates n number of control buses on the server

freeControlBus :: Int -> StateT Server IO ()

Frees a control bus, identified by number

allocAudioBus :: Int -> StateT Server IO (Maybe Int)

Allocates n number of audio buses on the server

freeAudioBus :: Int -> StateT Server IO ()

Frees an audio bus, identified by number

allocBuffer :: Int -> StateT Server IO (Maybe Int)

Allocates n number of buffers on the server

freeBuffer :: Int -> StateT Server IO ()

Frees a buffer, identified by number

sendDefaultGroup :: Server -> IO ()

Create the default group on the scsynth server. This is useful if you're starting it from the cml or otherwise not from sclang.

scprocessOut :: ServerProcess -> StateT Server IO ThreadId

Process the stdout and stderr from the running scsynth instance

quitMsg :: Message

A simple Osc Message that when sent will ask scsynth to quit

send :: Message -> StateT Server IO ()

Helper function, just wraps up the get and lift for slightly more streamlined sending. Used for things like set and free

sendBus :: Rate -> (Int -> Bus) -> StateT Server IO (Maybe Bus)

Takes a newBus function, allocates a bus, and returns a Bus object

sendNew :: ScSendable a => (Int -> a) -> StateT Server IO a

sendNew is a helper function for simplifying new node message sending to scsynth. It takes a function (NodeId to Node a), increments the node id in the server, and returns the created Node. Eg sendNew $ newSynth mySynth [(arg1, value)] ... or also: sendNew $ newGroup

sendSynth :: (Int -> Synth) -> StateT Server IO Synth

A Synth type based wrapper for sendNew

sendGroup :: (Int -> Group) -> StateT Server IO Group

A Group type based wrapper for sendNew

sendBuffer :: (Int -> Buffer) -> StateT Server IO (Maybe Buffer)

Similar to sendNew, this takes a curried function (returned by various Buffer module functions) and returns a Buffer When executed a buffer is allocated on the server, passed functions takes that number and returns a function which is sent to the server

loadDirectory :: String -> StateT Server IO ()

Requests scsynth to load a directory of compiled synth definitions

record :: StateT Server IO ()

Begin recording, this just uses the current working directory and a generated filename based on the current time.

pauseRecording :: StateT Server IO ()

Pauses the current recording, if any.

stopRecording :: StateT Server IO ()

Stops and finalized the current recording, if any.

prepareForRecording :: StateT Server IO ()

A helper function for created the necessary node and buffer for record. Not necessary to call this before recording.

boot :: StateT Server IO ()

Boots the server, spawning a child scsynth process and established pipes for stdout and sterr

quit :: StateT Server IO ()

Quits the child scsynth process

notify :: Message

A notify message, when notified scsynth will start sending update messages to sender regarding running synths, ugens, etc.

testSC :: IO ()

Test code for Haskollider. First start SC and boot the server. Next, you'll need a Synth named TestSine; Something like this:

SynthDef.new(TestSine, { arg freq = 440, amp = 0.2; var env = EnvGen.ar(Env.perc(0, 0.5), doneAction: 2); Out.ar(0, SinOsc.ar(freq, 0 , env * amp).dup); }).store;

Next, make sure you have Haskollider compiled and installed (cabal build && cabal install). Finally, open up ghci and run :m Haskollider.Server testSC

testSC2 :: IO ()

More test code for Haskollider. First start SC and boot the server. Next, you'll need a Synth named TestSine2; Something like this:

SynthDef.new(TestSine2, { arg freq = 440, amp = 0.2, gate = 1; var env = EnvGen.ar(Env.asr(0, 1, 0.5), gate: gate, doneAction: 2); Out.ar(0, SinOsc.ar(freq, 0 , env * amp).dup); }).store;

Next, make sure you have Haskollider compiled and installed (cabal build && cabal install). Finally, open up ghci and run :m Haskollider.Server testSC2

testSC3 :: IO ()

Server Process control test. Using the boot/quit functions we don't have to start SuperCollider outside of the program. This assumes you're on a unixy system with supercollider installed and with scsynth in your path. Also, for recording to work you need to have a server-record synth def like such compiled:

SynthDef(server-record, { arg bufnum; DiskOut.ar(bufnum, In.ar(0, recChannels)) }).store;