Haskollider.Util

Description

Utility functions for internal usage.

Documentation

flag :: Bool -> Int

Helper function to convert a Bool to Int for use in server Osc Messages

dropN :: (Enum b1, Eq b1, Num b1) => b1 -> [b] -> [b]

Drops m number of elements from a list

traceState :: Show a => a -> State s a

Print function for state monads

traceState2 :: Show s => a -> State s a

Print function for state monads, prints x

testState :: (Show a, Eq a) => a -> a -> State s a

Test print for state monads. Prints x == y

class ScSendable a where

Class for all sendable types, instanced by Synth, Group, Bus, and Buffer

Methods

newMsg :: a -> Message